requirements.txt
flask
flask-smorest
python-dotenv
sqlalchemy
flask_sqlalchemy
flask-jwt-extended
passlib
flask-migrate
gunicorn
psycopg2
requests
datetime
flask_apscheduler